requirements.txt
fastapi
uvicorn
matplotlib
PyPDF2
python-docx
scikit-learn
numpy
pandas
requests
python-multipart
